# docs/research/main_specs/1.0.0-mvp/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Helix OTA Control Plane API
  version: 1.0.0-MVP
  summary: REST API for the Helix OTA modular-monolith control plane (Android 15 first).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: >
    Control-plane endpoints (auth, devices, artifacts, releases, deployments, client
    update-check + telemetry). JSON responses negotiate Brotli/gzip; the referenced
    artifact-byte path is served identity-encoded, ZIP_STORED, with HTTP Range support.
    Response-header conventions (applied to every negotiated JSON response, not repeated per
    operation): 'Vary: Accept-Encoding' is always set (the Vary header component), and 'X-Request-Id'
    is always returned. Rate-limited routes additionally return RateLimit-Limit / RateLimit-Remaining
    / RateLimit-Reset (draft IETF RateLimit fields); on 429 a Retry-After header is returned.
  contact:
    name: Helix OTA control-plane team
  # license: intentionally omitted — project license is UNVERIFIED in this revision.
servers:
  - url: https://{host}/api/v1
    description: Helix OTA control plane (HTTP/3 primary, HTTP/2 fallback).
    variables:
      host:
        default: helix.example
...
security:
  - bearerAuth: []
paths:
...
  /artifacts/{artifactId}:
    get:
      tags: [artifacts]
      operationId: getArtifact
      summary: Artifact metadata (not the bytes; bytes are referenced from the update-check).
      parameters:
        - $ref: '#/components/parameters/ArtifactId'
      responses:
        '200':
          description: Artifact metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /deltas:
    post:
      tags: [deltas]
      operationId: registerDelta
      summary: Register a generated base->target delta artifact.
      description: >
        RBAC operator/admin. Both referenced artifacts MUST exist and MUST differ. A duplicate
        (base_artifact_id, target_artifact_id) pair is a 409 (handlers_delta.go:handleRegisterDelta).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeltaRegister' }
      responses:
        '201':
          description: Delta registered.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeltaView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [deltas]
      operationId: findDelta
      summary: Look up the delta for a base->target artifact pair.
      description: >
        RBAC viewer+. Returns the delta for ?base=<id>&target=<id>, or 404 if none exists
        (handlers_delta.go:handleFindDelta).
      parameters:
        - { name: base, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: target, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Delta found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeltaView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /branches:
    post:
      tags: [branches]
      operationId: createBranch
      summary: Create a release branch within a project.
      description: >
        RBAC operator/admin. A branch is a project-scoped release channel
        (handlers_branches.go:handleCreateBranch).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BranchCreate' }
      responses:
        '201':
          description: Branch created.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [branches]
      operationId: listBranches
      summary: List branches (account-scoped, cursor-paginated).
      description: RBAC viewer+ (handlers_branches.go:handleListBranches).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Branches list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /branches/{id}:
    get:
      tags: [branches]
      operationId: getBranch
      summary: Read a single branch.
      description: RBAC viewer+ (handlers_branches.go:handleGetBranch).
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: Branch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchView' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [branches]
      operationId: updateBranch
      summary: Update a branch.
      description: RBAC operator/admin (handlers_branches.go:handleUpdateBranch).
      parameters:
        - $ref: '#/components/parameters/BranchId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BranchUpdate' }
      responses:
        '200':
          description: Branch updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BranchView' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [branches]
      operationId: deleteBranch
      summary: Delete a branch.
      description: RBAC operator/admin (handlers_branches.go:handleDeleteBranch).
      parameters:
        - $ref: '#/components/parameters/BranchId'
      responses:
        '204':
          description: Branch deleted.
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --- Fabric registry (emulation test-fabric, admin-only) ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2/JWT access token (admin/operator/viewer) or device-scoped token (role=device).
        RBAC enforced per-route; device tokens may act only on their own deviceId.
    provisioningToken:
      type: http
      scheme: bearer
      description: >
        Alternative auth for POST /devices/register only: a provisioning token that authorizes a
        device to enrol itself when no operator bearer is available. The global bearerAuth (operator
        role) is the primary path; this scheme is the alternative (either satisfies the operation).
  headers:
    XRequestId:
      description: Correlation id; mirrors error.request_id.
      schema: { type: string }
...
  parameters:
...
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    BranchId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
  responses:
    ValidationFailed:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Missing, invalid, or expired credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but lacks the required role or resource ownership.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: State conflict (duplicate, non-monotonic version, overlapping deployment).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SemVer:
      type: string
      pattern: '^\d+\.\d+\.\d+$'
      examples: ['1.1.0']
    Os:
      type: string
      enum: [android]
      description: MVP targets Android 15 first; Linux/universal is a later phase.
    Sha256Hex:
      type: string
      pattern: '^[a-f0-9]{64}$'
      description: Lowercase-hex SHA-256.
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_FAILED
                - CONFLICT
                - UNSUPPORTED_MEDIA_TYPE
                - PAYLOAD_TOO_LARGE
                - RATE_LIMITED
                - SIGNATURE_INVALID
                - HASH_MISMATCH
                - VERSION_NOT_MONOTONIC
                - INTERNAL
            message: { type: string }
            request_id: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    Artifact:
      type: object
      required: [artifact_id, sha256, size, os, target_model, version, verified, uploaded_at]
      additionalProperties: false
      properties:
        artifact_id: { type: string, format: uuid }
        sha256: { $ref: '#/components/schemas/Sha256Hex' }
        size: { type: integer, format: int64 }
        os: { $ref: '#/components/schemas/Os' }
        target_model: { type: string }
        version: { $ref: '#/components/schemas/SemVer' }
        storage_ref: { type: string, description: Opaque object-store reference (MinIO/S3). }
        verified: { type: boolean }
        uploaded_at: { type: string, format: date-time }
...
    DeltaRegister:
      type: object
      required: [base_artifact_id, target_artifact_id]
      additionalProperties: false
      properties:
        base_artifact_id: { type: string, format: uuid }
        target_artifact_id: { type: string, format: uuid, description: Must differ from base_artifact_id. }
        sha256: { $ref: '#/components/schemas/Sha256Hex' }
        size: { type: integer, format: int64 }
        storage_ref: { type: string }
    DeltaView:
      type: object
      required: [id, base_artifact_id, target_artifact_id, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        base_artifact_id: { type: string, format: uuid }
        target_artifact_id: { type: string, format: uuid }
        sha256: { $ref: '#/components/schemas/Sha256Hex' }
        size: { type: integer, format: int64 }
        storage_ref: { type: string }
        created_at: { type: string, format: date-time }

    # --- staged rollout (handlers_rollout.go; ota-rollout-engine brick) ---
...
    BranchCreate:
      type: object
      required: [name, project_id]
      additionalProperties: false
      properties:
        name: { type: string }
        project_id: { type: string, format: uuid }
        description: { type: string }
    BranchUpdate:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        description: { type: string }
    BranchView:
      type: object
      required: [id, name, project_id, account_id, created_at]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        project_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BranchList:
      type: object
      required: [items]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BranchView' }
        next_cursor:
          oneOf:
            - { type: string }
            - { type: 'null' }

    # --- fabric registry (handlers_fabric.go) ---
...